Use CDF16 + unsigned short for NEE learnt distributions PDFs

+++ src/Renderer/CPUGPUCommonDataStructures/IlluminationAwareKDTreeDataHost.h
@@ -94,7 +94,7 @@ struct IlluminationAwareKDTreeDataHost
 		m_batch_spatial_moments	  = DataContainer<IlluminationAwareKDTreeSpatialSampleMoments>();
 		m_history_spatial_moments = DataContainer<IlluminationAwareKDTreeSpatialSampleMoments>();
 
-		m_tree_cut_sampling_probabilities = DataContainer<float>();
+		m_tree_cut_sampling_probabilities = DataContainer<unsigned short int>();
 		m_tree_cut_sampling_cdfs		  = DataContainer<unsigned short int>();
 
 		m_history_per_cell_sample_count				   = DataContainer<unsigned int>();
@@ -107,7 +107,7 @@ struct IlluminationAwareKDTreeDataHost
 		m_batch_per_cut_node_second_moment_sum		   = DataContainer<float>();
 		m_batch_per_cut_node_sample_count			   = DataContainer<unsigned int>();
 
-		m_tree_cut_sampling_prior_pdfs = DataContainer<float>();
+		m_tree_cut_sampling_prior_pdfs = DataContainer<unsigned short int>();
 		m_tree_cut_sampling_prior_cdfs = DataContainer<unsigned short int>();
 
 		return true;
@@ -199,7 +199,7 @@ struct IlluminationAwareKDTreeDataHost
 	DataContainer<IlluminationAwareKDTreeSpatialSampleMoments> m_history_spatial_moments;
 
 	// Buffers below that point are for learning NEE distributions per each guiding cell
-	DataContainer<float> m_tree_cut_sampling_probabilities;
+	DataContainer<unsigned short int> m_tree_cut_sampling_probabilities;
 	DataContainer<unsigned short int> m_tree_cut_sampling_cdfs;
 
 	DataContainer<GenericAtomicType<unsigned int, DataContainer>> m_history_per_cell_sample_count;
@@ -212,7 +212,7 @@ struct IlluminationAwareKDTreeDataHost
 	DataContainer<GenericAtomicType<float, DataContainer>> m_batch_per_cut_node_second_moment_sum;
 	DataContainer<GenericAtomicType<unsigned int, DataContainer>> m_batch_per_cut_node_sample_count;
 
-	DataContainer<float> m_tree_cut_sampling_prior_pdfs;
+	DataContainer<unsigned short int> m_tree_cut_sampling_prior_pdfs;
 	DataContainer<unsigned short int> m_tree_cut_sampling_prior_cdfs;
 };
